requirements.txt
autoenv==1.0.0
bcrypt==3.1.7
blessings==1.7
bpython==0.19
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
coverage==5.1
cryptography==2.9.2
curtsies==0.3.1
dnspython==1.16.0
ecdsa==0.15
email-validator==1.1.1
Fabric3==1.14.post1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.11
get==2019.4.13
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
invoke==1.4.1
itsdangerous==0.24
Jinja2==2.11.2
MarkupSafe==0.23
nose==1.3.4
odict==1.7.0
paramiko==2.7.1
Pillow==7.1.2
plumber==1.6
post==2019.4.13
psycopg2-binary==2.8.5
public==2019.4.13
pycparser==2.20
pycrypto==2.6.1
Pygments==2.6.1
PyNaCl==1.4.0
query-string==2019.4.13
request==2019.4.13
requests==2.24.0
six==1.15.0
SQLAlchemy==0.9.9
urllib3==1.25.9
wcwidth==0.2.4
Werkzeug==0.16.0
WTForms==2.0.2
zope.component==4.6.1
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.1.0
zope.lifecycleevent==4.3
zope.proxy==4.3.5
